+++ rest/admin/__tests__/2022-07/fulfillment_service.test.ts
@@ -66,7 +66,7 @@ describe('FulfillmentService resource', () => {
   });
 
   it('test_3', async () => {
-    queueMockResponse(JSON.stringify({"fulfillment_service": {"id": 1061774487, "name": "Jupiter Fulfillment", "email": null, "service_name": "Jupiter Fulfillment", "handle": "jupiter-fulfillment", "fulfillment_orders_opt_in": true, "include_pending_stock": false, "provider_id": null, "location_id": 1072404542, "callback_url": "http://google.com/", "tracking_support": true, "inventory_management": true, "admin_graphql_api_id": "gid://shopify/ApiFulfillmentService/1061774487", "permits_sku_sharing": true}}));
+    queueMockResponse(JSON.stringify({"fulfillment_service": {"id": 1061774524, "name": "Jupiter Fulfillment", "email": null, "service_name": "Jupiter Fulfillment", "handle": "jupiter-fulfillment", "fulfillment_orders_opt_in": true, "include_pending_stock": false, "provider_id": null, "location_id": 1072404579, "callback_url": "http://google.com/", "tracking_support": true, "inventory_management": true, "admin_graphql_api_id": "gid://shopify/ApiFulfillmentService/1061774524", "permits_sku_sharing": true}}));
 
     const fulfillment_service = new shopify.rest.FulfillmentService({session: session});
     fulfillment_service.name = "Jupiter Fulfillment";